+++ src/core/zowe/core_for_zowe_sdk/logger.py
@@ -1,12 +1,20 @@
+"""
+Logger module for handling application logging.
+
+This module provides the `Log` class which allows for registering,
+setting levels, opening, and closing loggers.
+"""
+
 import logging
 import os
 
 
 class Log:
-    """Class used to represent a logger
+    """
+    Class used to represent a logger.
 
     Attributes
-    -------
+    ----------
     loggers: set
         The set of all loggers
     dirname: str
@@ -28,27 +36,27 @@ class Log:
     file_handler.setFormatter(
         logging.Formatter("[%(asctime)s] [%(levelname)s] [%(name)s] - %(message)s", "%m/%d/%Y %I:%M:%S %p")
     )
-    console_handler = logging.StreamHandler()
+    console_handler: logging.StreamHandler = logging.StreamHandler()
 
     file_output: bool = True
     console_output: bool = True
 
-    loggers = set()
+    loggers: set = set()
 
     @staticmethod
-    def registerLogger(name: str):
+    def registerLogger(name: str) -> logging.Logger:
         """
-        Create and register a logger
+        Create and register a logger.
 
         Parameters
         ----------
         name: str
-            name for the logger
+            The name for the logger.
 
         Returns
         -------
-        Logger
-            A Logger object named after the file where it is created
+        logging.Logger
+            A Logger object named after the file where it is created.
         """
         logger = logging.getLogger(name)
         if Log.console_output:
@@ -62,7 +70,7 @@ class Log:
     @staticmethod
     def setAllLoggerLevel(level: int):
         """
-        Set display level for all loggers
+        Set display level for all loggers.
 
         Parameters
         ----------
@@ -75,25 +83,25 @@ class Log:
                 handler.setLevel(level)
 
     @staticmethod
-    def close(logger):
+    def close(logger: logging.Logger):
         """
-        Disable a logger
+        Disable a logger.
 
         Parameters
         ----------
-        logger: Logger
+        logger: logging.Logger
             The logger to be turned off
         """
         logger.disabled = True
 
     @staticmethod
-    def open(logger):
+    def open(logger: logging.Logger):
         """
-        Enable a logger
+        Enable a logger.
 
         Parameters
         ----------
-        logger: Logger
+        logger: logging.Logger
             The logger to be turned on
         """
         logger.disabled = False