+++ src/zos_files/zowe/zos_files_for_zowe_sdk/datasets.py
@@ -12,7 +12,7 @@ Copyright Contributors to the Zowe Project.
 
 import os
 from dataclasses import dataclass
-from typing import Optional
+from typing import Dict, List, Optional
 
 from zowe.core_for_zowe_sdk import SdkApi
 from zowe.core_for_zowe_sdk.exceptions import FileNotFound
@@ -23,41 +23,41 @@ _ZOWE_FILES_DEFAULT_ENCODING = zos_file_constants["ZoweFilesDefaultEncoding"]
 
 @dataclass
 class DatasetOption:
-    """A dataclass that represents options for creating a dataset
+    """A dataclass that represents options for creating a dataset.
 
-    Attributes
+    Parameters
     ----------
-    like: str
+    like : Optional[str]
         The dataset name to copy attributes from
-    volser: str
+    volser : Optional[str]
         The volume serial number that identifies where the dataset resides or should be allocated
-    unit: str
+    unit : Optional[str]
         Specifies the type of device on which the dataset is to be stored
-    dsorg: str
+    dsorg : Optional[str]
         Defines the organization of the dataset (PS for sequential, PO for partitioned, DA for direct access)
-    alcunit: str
+    alcunit : Optional[str]
         Specifies the unit of space allocation for the dataset (CYL for cylinders, TRK for tracks, BLK for blocks)
-    primary: int
+    primary : Optional[int]
         The amount of primary space to allocate for the dataset
-    secondary: int
+    secondary : Optional[int]
         The amount of secondary space to allocate if the primary space is exhausted
-    dirblk: int
+    dirblk : Optional[int]
         The number of directory blocks to allocate for a partitioned dataset
-    avgblk: int
+    avgblk : Optional[int]
         The average block size for the dataset
-    recfm: str
+    recfm : Optional[str]
         The format of the records in the dataset
-    blksize: int
+    blksize : Optional[int]
         The physical block size used for the dataset
-    lrecl: int
+    lrecl : Optional[int]
         The length of the logical records in the dataset
-    storclass: str
+    storclass : Optional[str]
         Specifies the storage class to be used for the dataset
-    mgnclass: str
-        Specifies the management class for the dataset
-    dataclass: str
+    mgmtclass : Optional[str]
+        Specifies the management class
+    dataclass : Optional[str]
         Specifies the data class for the dataset
-    dsntype: str
+    dsntype : Optional[str]
         Specifies the type of dataset
     """
 
@@ -76,7 +76,7 @@ class DatasetOption:
         blksize: Optional[int] = None,
         lrecl: Optional[int] = None,
         storclass: Optional[str] = None,
-        mgntclass: Optional[str] = None,
+        mgmtclass: Optional[str] = None,
         dataclass: Optional[str] = None,
         dsntype: Optional[str] = None,
     ) -> None:
@@ -93,42 +93,50 @@ class DatasetOption:
         self.lrecl = lrecl
         self.blksize = blksize
         self.storclass = storclass
-        self.mgntclass = mgntclass
+        self.mgmtclass = mgmtclass
         self.dataclass = dataclass
         self.dsntype = dsntype
 
     @property
     def volser(self) -> Optional[str]:
+        """Get the volume serial number."""
         return self.__volser
 
     @volser.setter
     def volser(self, volser: Optional[str]):
+        """Set the volume serial number."""
         self.__volser = volser
 
     @property
     def unit(self) -> Optional[str]:
+        """Get the type of device."""
         return self.__unit
 
     @unit.setter
     def unit(self, unit: Optional[str]):
+        """Set the type of device."""
         self.__unit = unit
 
     @property
     def dsorg(self) -> Optional[str]:
+        """Get the organization of the dataset."""
         return self.__dsorg
 
     @dsorg.setter
     def dsorg(self, dsorg: Optional[str]):
+        """Set the organization of the dataset."""
         if dsorg not in ("PO", "PS", None):
             raise ValueError("'dsorg' must be 'PO', 'PS', or None")
         self.__dsorg = dsorg
 
     @property
     def alcunit(self) -> Optional[str]:
+        """Get the unit of space allocation."""
         return self.__alcunit
 
     @alcunit.setter
     def alcunit(self, alcunit: Optional[str]):
+        """Set the unit of space allocation."""
         if alcunit is None:
             if self.like is not None:
                 self.__alcunit = None
@@ -141,10 +149,12 @@ class DatasetOption:
 
     @property
     def primary(self) -> Optional[int]:
+        """Get the primary space allocation."""
         return self.__primary
 
     @primary.setter
     def primary(self, primary: Optional[int]):
+        """Set the primary space allocation."""
         if primary is not None:
             if primary > 16777215:
                 raise ValueError("Given primary space allocation exceeds track limit of 16,777,215")
@@ -152,10 +162,12 @@ class DatasetOption:
 
     @property
     def secondary(self) -> Optional[int]:
+        """Get the secondary space allocation."""
         return self.__secondary
 
     @secondary.setter
     def secondary(self, secondary: Optional[int]):
+        """Set the secondary space allocation."""
         if self.primary is not None:
             secondary = secondary if secondary is not None else int(self.primary / 10)
             if secondary > 16777215:
@@ -164,26 +176,32 @@ class DatasetOption:
 
     @property
     def dirblk(self) -> Optional[int]:
+        """Get the number of directory blocks."""
         return self.__dirblk
 
     @dirblk.setter
     def dirblk(self, dirblk: Optional[int]):
+        """Set the number of directory blocks."""
         self.__dirblk = dirblk
 
     @property
     def avgblk(self) -> Optional[int]:
+        """Get the average block size."""
         return self.__avgblk
 
     @avgblk.setter
     def avgblk(self, avgblk: Optional[int]):
+        """Set the average block size."""
         self.__avgblk = avgblk
 
     @property
     def recfm(self) -> Optional[str]:
+        """Get the record format."""
         return self.__recfm
 
     @recfm.setter
     def recfm(self, recfm: Optional[str]):
+        """Set the record format."""
         if recfm is None:
             if self.like is not None:
                 self.__recfm = None
@@ -198,10 +216,12 @@ class DatasetOption:
 
     @property
     def blksize(self) -> Optional[int]:
+        """Get the physical block size."""
         return self.__blksize
 
     @blksize.setter
     def blksize(self, blksize: Optional[int]):
+        """Set the physical block size."""
         if blksize is None:
             if self.like is not None:
                 self.__blksize = None
@@ -213,96 +233,94 @@ class DatasetOption:
 
     @property
     def lrecl(self) -> Optional[int]:
+        """Get the length of logical records."""
         return self.__lrecl
 
     @lrecl.setter
     def lrecl(self, lrecl: Optional[int]):
+        """Set the length of logical records."""
         self.__lrecl = lrecl
 
     @property
     def storclass(self) -> Optional[str]:
+        """Get the storage class."""
         return self.__storclass
 
     @storclass.setter
     def storclass(self, storclass: Optional[str]):
+        """Set the storage class."""
         self.__storclass = storclass
 
     @property
-    def mgntclass(self) -> Optional[str]:
-        return self.__mgntclass
+    def mgmtclass(self) -> Optional[str]:
+        """Get the management class."""
+        return self.__mgmtclass
 
-    @mgntclass.setter
-    def mgntclass(self, mgntclass: Optional[str]):
-        self.__mgntclass = mgntclass
+    @mgmtclass.setter
+    def mgmtclass(self, mgmtclass: Optional[str]):
+        """Set the management class."""
+        self.__mgmtclass = mgmtclass
 
     @property
     def dataclass(self) -> Optional[str]:
+        """Get the data class."""
         return self.__dataclass
 
     @dataclass.setter
     def dataclass(self, dataclass: Optional[str]):
+        """Set the data class."""
         self.__dataclass = dataclass
 
     @property
     def dsntype(self) -> Optional[str]:
+        """Get the type of dataset."""
         return self.__dsntype
 
     @dsntype.setter
     def dsntype(self, dsntype: Optional[str]):
+        """Set the type of dataset."""
         self.__dsntype = dsntype
 
     @property
     def like(self) -> Optional[str]:
+        """Get the dataset name to copy attributes from."""
         return self.__like
 
     def to_dict(self) -> dict:
-        """Return the DatasetOption as a dict
-
-        Returns
-        -------
-        dict
-        """
+        """Return the DatasetOption as a dict."""
         return {key.replace("_DatasetOption__", ""): value for key, value in self.__dict__.items() if value is not None}
 
 
 class Datasets(SdkApi):
     """
-    Class used to represent the base z/OSMF Datasets API
-    which includes all operations related to datasets.
+    Class used to represent the base z/OSMF Datasets API.
 
-    Attributes
+    It includes all operations related to datasets.
+
+    Parameters
     ----------
-    connection
-        connection object
+    connection : dict
+        A profile for connection in dict (json) format
     """
 
-    def __init__(self, connection):
-        """
-        Construct a Datasets object.
-
-        Parameters
-        ----------
-        connection
-            The z/OSMF connection object (generated by the ZoweSDK object)
-
-        Also update header to accept gzip encoded responses
-        """
+    def __init__(self, connection: dict):
         super().__init__(connection, "/zosmf/restfiles/", logger_name=__name__)
         self._default_headers["Accept-Encoding"] = "gzip"
 
-    def list(self, name_pattern, return_attributes=False):
-        """Retrieve a list of datasets based on a given pattern.
+    def list(self, name_pattern: str, return_attributes: bool = False) -> List[Dict]:
+        """
+        Retrieve a list of datasets based on a given pattern.
 
         Parameters
         ----------
         name_pattern : str
             The pattern to match dataset names.
-        return_attributes : bool, optional
+        return_attributes : bool
             Whether to return dataset attributes along with the names. Defaults to False.
 
         Returns
         -------
-        list of dict
+        List[Dict]
             A JSON with a list of dataset names (and attributes if specified) matching the given pattern.
         """
         custom_args = self._create_custom_request_arguments()
@@ -315,25 +333,33 @@ class Datasets(SdkApi):
         response_json = self.request_handler.perform_request("GET", custom_args)
         return response_json
 
-    def list_members(self, dataset_name, member_pattern=None, member_start=None, limit=1000, attributes="member"):
-        """Retrieve the list of members on a given PDS/PDSE.
+    def list_members(
+        self,
+        dataset_name: str,
+        member_pattern: Optional[str] = None,
+        member_start: Optional[str] = None,
+        limit: int = 1000,
+        attributes: str = "member",
+    ) -> dict:
+        """
+        Retrieve the list of members on a given PDS/PDSE.
 
         Parameters
-        -------
+        ----------
         dataset_name: str
-            The name of the dataset
-        member_pattern: str
-            Filters members by name pattern.
-        member_start: str
-            The starting point for listing members.
+            Name of the dataset
+        member_pattern: Optional[str]
+            Filters members by name pattern
+        member_start: Optional[str]
+            The starting point for listing members
         limit: int
-            The maximum number of members returned.
+            The maximum number of members returned
         attributes: str
-            The member attributes to retrieve.
+            The member attributes to retrieve
 
         Returns
         -------
-        json
+        dict
             A JSON with a list of members from a given PDS/PDSE
         """
         custom_args = self._create_custom_request_arguments()
@@ -351,43 +377,47 @@ class Datasets(SdkApi):
 
     def copy_data_set_or_member(
         self,
-        from_dataset_name,
-        to_dataset_name,
-        from_member_name=None,
-        volser=None,
-        alias=None,
-        to_member_name=None,
-        enq=None,
-        replace=False,
-    ):
+        from_dataset_name: str,
+        to_dataset_name: str,
+        from_member_name: Optional[str] = None,
+        volser: Optional[str] = None,
+        alias: Optional[bool] = None,
+        to_member_name: Optional[str] = None,
+        enq: Optional[str] = None,
+        replace: bool = False,
+    ) -> dict:
         """
         Copy a dataset or member to another dataset or member.
 
         Parameters
-        -------
+        ----------
         from_dataset_name: str
             Name of the dataset to copy from
         to_dataset_name: str
             Name of the dataset to copy to
-        from_member_name: str
+        from_member_name: Optional[str]
             Name of the member to copy from
-        volser: str
+        volser: Optional[str]
             Volume serial number of the dataset to copy from
-        alias: bool
+        alias: Optional[bool]
             Alias of the dataset to copy from
-        to_member_name: str
+        to_member_name: Optional[str]
             Name of the member to copy to
-        enq: str
+        enq: Optional[str]
             Enqueue type for the dataset to copy from
         replace: bool
-            If true, members in the target data set are replaced.
+            If true, members in the target data set are replaced
+
+        Raises
+        ------
+        ValueError
+            Thrown when enq has an invalid value
 
         Returns
         -------
-        json
+        dict
             A JSON containing the result of the operation
         """
-
         data = {
             "request": "copy",
             "from-dataset": {"dsn": from_dataset_name.strip(), "member": from_member_name},
@@ -412,7 +442,7 @@ class Datasets(SdkApi):
         response_json = self.request_handler.perform_request("PUT", custom_args, expected_code=[200])
         return response_json
 
-    def create(self, dataset_name, options: Optional[DatasetOption] = None):
+    def create(self, dataset_name: str, options: Optional[DatasetOption] = None) -> dict:
         """
         Create a sequential or partitioned dataset.
 
@@ -420,12 +450,17 @@ class Datasets(SdkApi):
         ----------
         dataset_name: str
             Name of the dataset to be created
-        options: DatasetOption
+        options: Optional[DatasetOption]
             A DatasetOption class with property options of the dataset
+
+        Raises
+        ------
+        ValueError
+            Thrown when a parameter has an invalid value
 
         Returns
         -------
-        json
+        dict
             A JSON containing the result of the operation
         """
         if not options:
@@ -458,9 +493,10 @@ class Datasets(SdkApi):
         response_json = self.request_handler.perform_request("POST", custom_args, expected_code=[201])
         return response_json
 
-    def create_default(self, dataset_name: str, default_type: str):
+    def create_default(self, dataset_name: str, default_type: str) -> dict:
         """
         Create a dataset with default options set.
+
         Default options depend on the requested type.
 
         Parameters
@@ -470,12 +506,16 @@ class Datasets(SdkApi):
         default_type: str
             The type of the dataset: "partitioned", "sequential", "classic", "c" or "binary"
 
+        Raises
+        ------
+        ValueError
+            Thrown when a parameter is invalid
+
         Returns
         -------
-        json
+        dict
             A JSON containing the result of the operation
         """
-
         if default_type not in ("partitioned", "sequential", "classic", "c", "binary"):
             self.logger.error("Invalid type for default data set.")
             raise ValueError("Invalid type for default data set.")
@@ -536,19 +576,20 @@ class Datasets(SdkApi):
         response_json = self.request_handler.perform_request("POST", custom_args, expected_code=[201])
         return response_json
 
-    def get_content(self, dataset_name, stream=False):
-        """Retrieve the contents of a given dataset.
+    def get_content(self, dataset_name: str, stream: bool = False) -> dict:
+        """
+        Retrieve the contents of a given dataset.
 
         Parameters
-        -------
+        ----------
         dataset_name: str
             The name of the dataset
         stream: bool
-            Specifies
+            Specifies whether the response is streamed. Default: False
 
         Returns
         -------
-        json
+        dict
             A JSON with the contents of a given dataset
         """
         custom_args = self._create_custom_request_arguments()
@@ -556,7 +597,7 @@ class Datasets(SdkApi):
         response_json = self.request_handler.perform_request("GET", custom_args, stream=stream)
         return response_json
 
-    def get_binary_content(self, dataset_name, stream=False, with_prefixes=False):
+    def get_binary_content(self, dataset_name: str, stream: bool = False, with_prefixes: bool = False) -> dict:
         """
         Retrieve the contents of a given dataset as a binary bytes object.
 
@@ -564,13 +605,15 @@ class Datasets(SdkApi):
         ----------
         dataset_name: str
             Name of the dataset to retrieve
-        with_prefixes: boolean
+        stream: bool
+            Specifies whether the request is streaming
+        with_prefixes: bool
             If True include a 4 byte big endian record len prefix. Default: False
 
         Returns
         -------
-        response
-            A response object from the requests library
+        dict
+            A JSON with the contents of a given dataset
         """
         custom_args = self._create_custom_request_arguments()
         custom_args["url"] = "{}ds/{}".format(self._request_endpoint, self._encode_uri_component(dataset_name))
@@ -582,8 +625,9 @@ class Datasets(SdkApi):
         response = self.request_handler.perform_request("GET", custom_args, stream=stream)
         return response
 
-    def write(self, dataset_name, data, encoding=_ZOWE_FILES_DEFAULT_ENCODING):
-        """Write content to an existing dataset.
+    def write(self, dataset_name: str, data: str, encoding: str = _ZOWE_FILES_DEFAULT_ENCODING) -> dict:
+        """
+        Write content to an existing dataset.
 
         Parameters
         ----------
@@ -591,12 +635,12 @@ class Datasets(SdkApi):
             Name of the dataset to retrieve
         data: str
             Content to be written
-        encoding:
-            Specifies encoding schema
+        encoding: str
+            Specifies encoding name (e.g. IBM-1047)
 
         Returns
         -------
-        json
+        dict
             A JSON containing the result of the operation
         """
         custom_args = self._create_custom_request_arguments()
@@ -606,11 +650,12 @@ class Datasets(SdkApi):
         response_json = self.request_handler.perform_request("PUT", custom_args, expected_code=[204, 201])
         return response_json
 
-    def download(self, dataset_name, output_file):
-        """Retrieve the contents of a dataset and saves it to a given file.
+    def download(self, dataset_name: str, output_file: str) -> None:
+        """
+        Retrieve the contents of a dataset and saves it to a given file.
 
         Parameters
-        -------
+        ----------
         dataset_name: str
             Name of the dataset to be downloaded
         output_file: str
@@ -621,36 +666,45 @@ class Datasets(SdkApi):
             for chunk in response.iter_content(chunk_size=4096, decode_unicode=True):
                 f.write(chunk)
 
-    def download_binary(self, dataset_name, output_file, with_prefixes=False):
-        """Retrieve the contents of a binary dataset and saves it to a given file.
+    def download_binary(self, dataset_name: str, output_file: str, with_prefixes: bool = False) -> None:
+        """
+        Retrieve the contents of a binary dataset and saves it to a given file.
 
         Parameters
         ----------
-        dataset_name:str
+        dataset_name : str
             Name of the dataset to download
-        output_file:str
+        output_file : str
             Name of the local file to create
-        with_prefixes:boolean
-            If true, include a four big endian bytes record length prefix. The default is False
+        with_prefixes : bool
+            If true, include a four big endian bytes record length prefix. Default: False
         """
         response = self.get_binary_content(dataset_name, with_prefixes=with_prefixes, stream=True)
         with open(output_file, "wb") as f:
             for chunk in response.iter_content(chunk_size=4096):
                 f.write(chunk)
 
-    def upload_file(self, input_file, dataset_name, encoding=_ZOWE_FILES_DEFAULT_ENCODING, binary=False):
-        """Upload contents of a given file and uploads it to a dataset.
+    def upload_file(
+        self, input_file: str, dataset_name: str, encoding: str = _ZOWE_FILES_DEFAULT_ENCODING, binary: bool = False
+    ) -> None:
+        """
+        Upload contents of a given file and uploads it to a dataset.
 
         Parameters
-        -------
+        ----------
         input_file: str
             Name of the file to be uploaded
         dataset_name: str
             Name of the dataset to be created
         encoding: str
-            Specifies the encoding schema
+            Specifies the encoding name (e.g. IBM-1047)
         binary: bool
             specifies whether the file is binary
+
+        Raises
+        ------
+        FileNotFound
+            Thrown when a file is not found at provided location
         """
         if os.path.isfile(input_file):
             if binary:
@@ -663,24 +717,22 @@ class Datasets(SdkApi):
             self.logger.error(f"File {input_file} not found.")
             raise FileNotFound(input_file)
 
-    def recall_migrated(self, dataset_name: str, wait=False):
+    def recall_migrated(self, dataset_name: str, wait: bool = False) -> dict:
         """
-        Recalls a migrated data set.
+        Recall a migrated data set.
 
         Parameters
         ----------
         dataset_name: str
             Name of the data set
-
         wait: bool
             If true, the function waits for completion of the request, otherwise the request is queued
 
         Returns
         -------
-        json
+        dict
             A JSON containing the result of the operation
         """
-
         data = {"request": "hrecall", "wait": wait}
 
         custom_args = self._create_custom_request_arguments()
@@ -690,27 +742,24 @@ class Datasets(SdkApi):
         response_json = self.request_handler.perform_request("PUT", custom_args, expected_code=[200])
         return response_json
 
-    def delete_migrated(self, dataset_name: str, purge=False, wait=False):
+    def delete_migrated(self, dataset_name: str, purge: bool = False, wait: bool = False) -> dict:
         """
-        Deletes migrated data set.
+        Delete migrated data set.
 
         Parameters
         ----------
         dataset_name: str
             Name of the data set
-
         purge: bool
             If true, the function uses the PURGE=YES on ARCHDEL request, otherwise it uses the PURGE=NO.
-
         wait: bool
             If true, the function waits for completion of the request, otherwise the request is queued.
 
         Returns
         -------
-        json
+        dict
             A JSON containing the result of the operation
         """
-
         data = {
             "request": "hdelete",
             "purge": purge,
@@ -724,24 +773,22 @@ class Datasets(SdkApi):
         response_json = self.request_handler.perform_request("PUT", custom_args, expected_code=[200])
         return response_json
 
-    def migrate(self, dataset_name: str, wait=False):
+    def migrate(self, dataset_name: str, wait: bool = False) -> dict:
         """
-        Migrates the data set.
+        Migrate the data set.
 
         Parameters
         ----------
         dataset_name: str
             Name of the data set
-
         wait: bool
             If true, the function waits for completion of the request, otherwise the request is queued.
 
         Returns
         -------
-        json
+        dict
             A JSON containing the result of the operation
         """
-
         data = {"request": "hmigrate", "wait": wait}
 
         custom_args = self._create_custom_request_arguments()
@@ -751,9 +798,9 @@ class Datasets(SdkApi):
         response_json = self.request_handler.perform_request("PUT", custom_args, expected_code=[200])
         return response_json
 
-    def rename(self, before_dataset_name: str, after_dataset_name: str):
+    def rename(self, before_dataset_name: str, after_dataset_name: str) -> dict:
         """
-        Renames the data set.
+        Rename the data set.
 
         Parameters
         ----------
@@ -765,10 +812,9 @@ class Datasets(SdkApi):
 
         Returns
         -------
-        json
+        dict
             A JSON containing the result of the operation
         """
-
         data = {"request": "rename", "from-dataset": {"dsn": before_dataset_name.strip()}}
 
         custom_args = self._create_custom_request_arguments()
@@ -780,30 +826,31 @@ class Datasets(SdkApi):
         response_json = self.request_handler.perform_request("PUT", custom_args, expected_code=[200])
         return response_json
 
-    def rename_member(self, dataset_name: str, before_member_name: str, after_member_name: str, enq=""):
+    def rename_member(self, dataset_name: str, before_member_name: str, after_member_name: str, enq: str = "") -> dict:
         """
-        Renames the data set member.
+        Rename the data set member.
 
         Parameters
         ----------
         dataset_name: str
             Name of the data set.
-
         before_member_name: str
             The source member name.
-
         after_member_name: str
             New name for the source member.
-
         enq: str
             Values can be SHRW or EXCLU. SHRW is the default for PDS members, EXCLU otherwise.
 
+        Raises
+        ------
+        ValueError
+            Thrown when a parameter is invalid
+
         Returns
         -------
-        json
+        dict
             A JSON containing the result of the operation
         """
-
         data = {
             "request": "rename",
             "from-dataset": {
@@ -828,21 +875,22 @@ class Datasets(SdkApi):
         response_json = self.request_handler.perform_request("PUT", custom_args, expected_code=[200])
         return response_json
 
-    def delete(self, dataset_name, volume=None, member_name=None):
-        """Deletes a sequential or partitioned data.
+    def delete(self, dataset_name: str, volume: Optional[str] = None, member_name: Optional[str] = None) -> dict:
+        """
+        Delete a sequential or partitioned data.
 
         Parameters
         ----------
         dataset_name: str
             The name of the dataset
-        volume: str
+        volume: Optional[str]
             The optional volume serial number
-        member_name: str
+        member_name: Optional[str]
             The name of the member to be deleted
 
         Returns
         -------
-        json
+        dict
             A JSON containing the result of the operation
         """
         custom_args = self._create_custom_request_arguments()
@@ -856,8 +904,13 @@ class Datasets(SdkApi):
         return response_json
 
     def copy_uss_to_data_set(
-        self, from_filename, to_dataset_name, to_member_name=None, type=FileType.TEXT, replace=False
-    ):
+        self,
+        from_filename: str,
+        to_dataset_name: str,
+        to_member_name: Optional[str] = None,
+        type: str = FileType.TEXT,
+        replace: bool = False,
+    ) -> dict:
         """
         Copy a USS file to dataset.
 
@@ -867,19 +920,18 @@ class Datasets(SdkApi):
             Name of the file to copy from.
         to_dataset_name: str
             Name of the dataset to copy to.
-        to_member_name: str
+        to_member_name: Optional[str]
             Name of the member to copy to.
-        type: FileType, optional
+        type: str
             Type of the file to copy from. Default is FileType.TEXT.
-        replace: bool, optional
+        replace: bool
             If true, members in the target dataset are replaced.
 
         Returns
         -------
-        json
+        dict
             A JSON containing the result of the operation.
         """
-
         data = {
             "request": "copy",
             "from-file": {"filename": from_filename.strip(), "type": type.value},